Refactored _updateObject function into new item helper class, and update all formData references to new object structure
