+++ modules/actors/actor-sheet-ffg.js
@@ -353,7 +353,10 @@ export class ActorSheetFFG extends ActorSheet {
 
               CONFIG.logger.debug(`Updating ${ability} Characteristic from ${characteristic} to ${newCharacteristic}`);
 
-              this.object.update({ [`data.skills.${ability}.characteristic`]: newCharacteristic });
+              let updateData;
+              setProperty(updateData, `data.skills.${ability}.characteristic`, newCharacteristic);
+
+              this.object.update(updateData);
             },
           },
           two: {
@@ -399,8 +402,10 @@ export class ActorSheetFFG extends ActorSheet {
 
               if (name.trim().length > 0) {
                 CONFIG.logger.debug(`Creating new skill ${name} (${characteristic})`);
+                let updateData;
+                setProperty(updateData, `data.skills.${name}`, newSkill);
 
-                this.object.update({ [`data.skills.${name}`]: newSkill });
+                this.object.update(updateData);
               }
             },
           },
